Extract concurrency timeline recorder from diagnostics controller
Keep the controller under the 1k-line bar by owning timeline append, finish sink, and gauge sampling in one recorder with a single record() API, and move event/occupancy wire helpers out of rollup.
