+++ packages/api/api/compute/diagnostics/rollup.py
@@ -142,46 +142,3 @@ def rollup_to_wire(rollup: ConcurrencyTimelineRollup) -> dict[str, Any]:
         "maxGlobalInFlight": rollup.max_global_in_flight,
         "configuredWorkers": rollup.configured_workers,
     }
-
-
-def event_to_wire(event: ComputeConcurrencyEvent) -> dict[str, Any]:
-    """CamelCase wire shape for one concurrency timeline event."""
-    return {
-        "kind": event.kind,
-        "timestamp": event.timestamp,
-        "scopeKey": event.scope_key,
-        "stepKind": event.step_kind,
-        "stepIndex": event.step_index,
-        "priorityBand": event.priority_band,
-        "backend": event.backend,
-        "executionKey": event.execution_key,
-        "terminalState": event.terminal_state,
-        "durationMs": event.duration_ms,
-        "gauges": {
-            "scopedReadyDepth": event.gauges.scoped_ready_depth,
-            "scopedInFlightCount": event.gauges.scoped_in_flight_count,
-            "globalInFlightCount": event.gauges.global_in_flight_count,
-            "globalQueueDepth": event.gauges.global_queue_depth,
-            "configuredWorkers": event.gauges.configured_workers,
-        },
-    }
-
-
-def live_occupancy_to_wire(
-    *,
-    configured_workers: int,
-    scoped_ready_depth: int,
-    scoped_in_flight_count: int,
-    global_in_flight_count: int,
-    global_queue_depth: int,
-    backend_mix: dict[str, int],
-) -> dict[str, Any]:
-    """CamelCase live occupancy snapshot fields."""
-    return {
-        "configuredWorkers": configured_workers,
-        "scopedReadyDepth": scoped_ready_depth,
-        "scopedInFlightCount": scoped_in_flight_count,
-        "globalInFlightCount": global_in_flight_count,
-        "globalQueueDepth": global_queue_depth,
-        "backendMix": backend_mix,
-    }
